# Preemption overhead bug fix